Updated to 1.5.0
Added GUI for phaser drill. Added Klingon NPC. Added duranium and
dilithium blocks. Retooled EntityThrowable into EntityProjectile.
