+++ main/java/startrekmod/graphics/RenderPhaserBlast.java
@@ -20,7 +20,7 @@ public class RenderPhaserBlast extends Render
 		GL11.glTranslated(transX, transY, transZ);
 		GL11.glScaled(1, 1, 1);
 		//rotates around vector specified by 3 values
-		GL11.glRotated(entity.rotationYaw + 90, 0, 1, 0);
+		GL11.glRotated(entity.rotationYaw + 90, 0, -1, 0);
 		//first call to glRotated rotates complete coordinate set
 		//no need for trigonometric re-calculation of axes
 		GL11.glRotated(entity.rotationPitch, 0, 0, -1);
