+++ main/java/startrekmod/network/handler/HandlerPhaserDrill.java
@@ -2,11 +2,10 @@ package startrekmod.network.handler;
 
 import startrekmod.network.packet.PacketPhaserDrill;
 import net.minecraft.entity.Entity;
+import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.world.World;
-
 import startrekmod.entity.EntityPhaserDrill;
 import startrekmod.network.packet.*;
-
 import cpw.mods.fml.common.network.simpleimpl.*;
 
 public class HandlerPhaserDrill implements IMessageHandler<PacketPhaserDrill, IMessage>
@@ -22,6 +21,7 @@ public class HandlerPhaserDrill implements IMessageHandler<PacketPhaserDrill, IM
 		if(!(entity instanceof EntityPhaserDrill)) return null;
 		
 		EntityPhaserDrill drill = (EntityPhaserDrill)entity;
+		drill.operator = (EntityPlayer)world.getEntityByID(message.playerID);
 		
 		if(message.direction == null)
 			drill.countdownTicks = 100;
