+++ main/java/startrekmod/network/packet/PacketPhaserDrill.java
@@ -1,5 +1,6 @@
 package startrekmod.network.packet;
 
+import net.minecraft.entity.player.EntityPlayer;
 import startrekmod.entity.EntityPhaserDrill;
 import startrekmod.util.DirectionMode;
 import io.netty.buffer.ByteBuf;
@@ -9,13 +10,15 @@ public class PacketPhaserDrill implements IMessage
 {
 	public DirectionMode direction;
 	public int drillID;
+	public int playerID;
 	
 	public PacketPhaserDrill() {} //required for processing
 	
-	public PacketPhaserDrill(DirectionMode direction, EntityPhaserDrill drill)
+	public PacketPhaserDrill(DirectionMode direction, EntityPhaserDrill drill, EntityPlayer player)
 	{
 		this.direction = direction;
 		drillID = drill.getEntityId();
+		playerID = player.getEntityId();
 	}
 
 	@Override
@@ -24,6 +27,7 @@ public class PacketPhaserDrill implements IMessage
 		int angle = data.readInt();
 		direction = (angle != 1) ? DirectionMode.fromAngle(angle) : null;
 		drillID = data.readInt();
+		playerID = data.readInt();
 	}
 
 	@Override
@@ -31,5 +35,6 @@ public class PacketPhaserDrill implements IMessage
 	{
 		data.writeInt((direction != null) ? direction.angle : 1);
 		data.writeInt(drillID);
+		data.writeInt(playerID);
 	}
 }
